Add length argument and punctuation to password command
This commit changes the !password command to use a wider range of
characters when creating the output password: lower and upper case ascii
characters, digits, and punctuation characters.

It also allows a parameter to be passed to the command to specify the
length of the password (default 12; see
https://blog.codinghorror.com/your-password-is-too-damn-short/).

If the length provided is greater than or equal to 3, the password is
guaranteed to include at least one of each type of character.